Change file organization.
